web/tstr-automation/requirements.txt
beautifulsoup4==4.12.3
requests==2.31.0
schedule==1.2.0
supabase==2.0.0
python-dotenv==1.0.0
functions-framework==3.5.0
flask==3.0.0
psycopg2-binary==2.9.11
